pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rapumamd"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["click", "pyyaml"]

[project.scripts]
rapumamd = "rapumamd.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
